+++ delete/variant.php
@@ -16,6 +16,14 @@ $event = "Deletion";
 $id = htmlentities($_POST['id']);
 $date = htmlentities($_POST['date']);
 $test = htmlentities($_POST['test']);
+$gene = htmlentities($_POST['gene']);
+$cdna = htmlentities($_POST['cdna']);
+$protein = htmlentities($_POST['protein']);
+$mutationid = htmlentities($_POST['mutationid']);
+$mutationhgvs = htmlentities($_POST['mutationhgvs']);
+$interpretation = htmlentities($_POST['interpretation']);
+$source = htmlentities($_POST['source']);
+$comment = str_replace("'", "\'", htmlentities($_POST['comment']));
 
 //Encryption
 $encryption_key = hex2bin($key);
@@ -31,7 +39,24 @@ $enc_id = bin2hex(openssl_encrypt($id, $cipher, $encryption_key, 0, $iv));
 
 $hasAdminRole = in_array("admin", explode(",", strtolower($roles)));
 
-if ($hasAdminRole) {
+$canDelete = false;
+$sql = "SELECT * FROM `deletion_tracking` WHERE username = ?";
+$stmt = $clinical_data_pdo->prepare($sql);
+$stmt->bindParam(1, $username);
+$checkCanDelete = $stmt->execute();
+
+if ($checkCanDelete && $stmt->rowCount() > 0) {
+    $now = date(DATETIME_FORMAT);
+    $row = $stmt->fetch();
+    $lastDeletion = strtotime($row['last_deletion']);
+
+    $difference = abs(strtotime($now) - $lastDeletion);
+    $canDelete = $difference > 86400; // we can delete only if the difference is greater than 24 hours in seconds
+} elseif ($stmt->rowCount() == 0) {
+    $canDelete = true;
+}
+
+if ($hasAdminRole && $canDelete) {
     $sql = "DELETE FROM `Variant`
         WHERE
             `id` = UNHEX(?)
@@ -98,6 +123,20 @@ if ($hasAdminRole) {
     $auditResult = $stmt3->execute();
 
     if ($mainResult && $trackingResult && $auditResult) {
+        $now = date(DATETIME_FORMAT);
+        $sql = "
+            INSERT INTO `deletion_tracking` (
+                `username`,
+                `last_deletion`
+            )
+            VALUES (?, ?)
+            ON DUPLICATE KEY UPDATE `last_deletion`=?
+        ";
+        $stmt = $clinical_data_pdo->prepare($sql);
+        $stmt->bindParam(1, $username);
+        $stmt->bindParam(2, $now);
+        $stmt->bindParam(3, $now);
+        $stmt->execute();
         echo "Success";
     } else {
         $error = null;
@@ -111,6 +150,10 @@ if ($hasAdminRole) {
         echo "There was a problem while deleting the data. ";
         echo "Please contact the admin of the site - Nadia Znassi. Your reference: " . $tracking . ":" . $error;
     }
+} elseif (!$canDelete) {
+    echo "You cannot delete data until 24 hours have passed since the last time you deleted data!";
+} else {
+    echo "You are not authorized to do this operation!";
 }
 
 mysqli_close($conn);
